+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## unreleased
+## 0.4.0 - 2024-03-15
 ### Added
 * Added `OutputAssignments`, `OutputAssignment`, `ChannelAssignment`, and `ChanelAssignmentTexture`.
 * Added method `output_assignments` to `Material` for accessing shader JSON database information.
 
 ### Changed
+* Update xc3_model to 0.7.0.
 * Changed the types of `MaterialParameters` fields to match xc3_model changes.
 * Renamed `Material.unk_type` to `Material.pass_type` to match xc3_model.